feat: 공지사항 상세페이지 메뉴클릭 hooks 추가
